change: Colocando as funções show de lawyer e appointment aos seus respectivos usecases
