+++ src/modules/appointment/usecases.h
@@ -1,0 +1,1 @@
+void showAppointmentMenu(void);
