+++ src/modules/lawyer/usecases.h
@@ -1,0 +1,1 @@
+void showLawyerMenu(void);
