+++ src/utils/interfaces.h
@@ -29,11 +29,7 @@ void showMainMenu(void);
 
 void showClientMenu(void);
 
-void showLawyerMenu(void);
-
 void showOfficeMenu(void);
-
-void showAppointmentMenu(void);
 
 void showAboutMenu(void);
